@@ -1,21 +1,34 @@
 // protoc plugin which converts .proto to JSON schema
-// It is spawned by protoc and generates JSON-schema files.
+// It is called by protoc and generates JSON-schema files.
 // "Heavily influenced" by Google's "protog-gen-bq-schema"
 //
 // usage:
-//  $ bin/protoc --jsonschema_out=path/to/outdir foo.proto
 //
+//	$ bin/protoc --jsonschema_out=path/to/outdir foo.proto
 package main
 
 import (
+	"flag"
 	"fmt"
 	"os"
 
-	"github.com/golang/protobuf/proto"
-	plugin "github.com/golang/protobuf/protoc-gen-go/plugin"
 	"github.com/sirupsen/logrus"
-	"github.com/sixt/protoc-gen-jsonschema/internal/converter"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/pluginpb"
+
+	"github.com/chrusty/protoc-gen-jsonschema/internal/converter"
 )
+
+const version = "v1.4.0"
+
+func init() {
+	versionFlag := flag.Bool("version", false, "prints current version")
+	flag.Parse()
+	if *versionFlag {
+		fmt.Println(version)
+		os.Exit(0)
+	}
+}
 
 func main() {
 
@@ -28,15 +41,14 @@ func main() {
 	protoConverter := converter.New(logger)
 
 	// Convert the generator request:
-	var ok = true
+	var failed bool
 	logger.Debug("Processing code generator request")
 	res, err := protoConverter.ConvertFrom(os.Stdin)
 	if err != nil {
-		ok = false
+		failed = true
 		if res == nil {
-			message := fmt.Sprintf("Failed to read input: %v", err)
-			res = &plugin.CodeGeneratorResponse{
-				Error: &message,
+			res = &pluginpb.CodeGeneratorResponse{
+				Error: proto.String(fmt.Sprintf("Failed to read input: %v", err)),
 			}
 		}
 	}
@@ -51,10 +63,10 @@ func main() {
 		logger.WithError(err).Fatal("Failed to write response")
 	}
 
-	if ok {
-		logger.Debug("Succeeded to process code generator request")
-	} else {
+	if failed {
 		logger.Warn("Failed to process code generator but successfully sent the error to protoc")
 		os.Exit(1)
 	}
+
+	logger.Debug("Succeeded to process code generator request")
 }
