@@ -1,7 +1,7 @@
 package testdata
 
 const ImportedEnum = `{
-    "$schema": "http://json-schema.org/draft-04/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "enum": [
         "VALUE_0",
         0,
@@ -14,10 +14,42 @@ const ImportedEnum = `{
     ],
     "oneOf": [
         {
-            "type": "string"
+            "description": "Zero",
+            "const": "VALUE_0"
         },
         {
-            "type": "integer"
+            "description": "Zero",
+            "const": 0
+        },
+        {
+            "description": "One",
+            "const": "VALUE_1"
+        },
+        {
+            "description": "One",
+            "const": 1
+        },
+        {
+            "description": "Two",
+            "const": "VALUE_2"
+        },
+        {
+            "description": "Two",
+            "const": 2
+        },
+        {
+            "description": "Three",
+            "const": "VALUE_3"
+        },
+        {
+            "description": "Three",
+            "const": 3
         }
-    ]
+    ],
+    "title": "Imported Enum",
+    "description": "This is an enum"
 }`
+
+const ImportedEnumFail = `"VALUE_5"`
+
+const ImportedEnumPass = `"VALUE_3"`
