@@ -2,8 +2,23 @@ package testdata
 
 const NestedMessage = `{
     "$schema": "http://json-schema.org/draft-04/schema#",
-    "properties": {
-        "payload": {
+    "$ref": "#/definitions/NestedMessage",
+    "definitions": {
+        "NestedMessage": {
+            "properties": {
+                "payload": {
+                    "$ref": "#/definitions/samples.PayloadMessage",
+                    "additionalProperties": true
+                },
+                "description": {
+                    "type": "string"
+                }
+            },
+            "additionalProperties": true,
+            "type": "object",
+            "title": "Nested Message"
+        },
+        "samples.PayloadMessage": {
             "properties": {
                 "name": {
                     "type": "string"
@@ -42,16 +57,25 @@ const NestedMessage = `{
                         {
                             "type": "integer"
                         }
-                    ]
+                    ],
+                    "title": "Topology"
                 }
             },
             "additionalProperties": true,
-            "type": "object"
-        },
-        "description": {
-            "type": "string"
+            "type": "object",
+            "title": "Payload Message"
         }
-    },
-    "additionalProperties": true,
-    "type": "object"
+    }
+}`
+
+const NestedMessageFail = `{
+	"payload": {
+		"topology": "ROUND"
+	}
+}`
+
+const NestedMessagePass = `{
+	"payload": {
+		"topology": "FLAT"
+	}
 }`
