@@ -19,5 +19,10 @@ const SecondEnum = `{
         {
             "type": "integer"
         }
-    ]
+    ],
+    "title": "Second Enum"
 }`
+
+const SecondEnumFail = `"VALUE_3"`
+
+const SecondEnumPass = `"VALUE_7"`
