@@ -2,23 +2,33 @@ package testdata
 
 const SecondMessage = `{
     "$schema": "http://json-schema.org/draft-04/schema#",
-    "properties": {
-        "name2": {
-            "type": "string"
-        },
-        "timestamp2": {
-            "type": "string"
-        },
-        "id2": {
-            "type": "integer"
-        },
-        "rating2": {
-            "type": "number"
-        },
-        "complete2": {
-            "type": "boolean"
+    "$ref": "#/definitions/SecondMessage",
+    "definitions": {
+        "SecondMessage": {
+            "properties": {
+                "name2": {
+                    "type": "string"
+                },
+                "timestamp2": {
+                    "type": "string"
+                },
+                "id2": {
+                    "type": "integer"
+                },
+                "rating2": {
+                    "type": "number"
+                },
+                "complete2": {
+                    "type": "boolean"
+                }
+            },
+            "additionalProperties": true,
+            "type": "object",
+            "title": "Second Message"
         }
-    },
-    "additionalProperties": true,
-    "type": "object"
+    }
 }`
+
+const SecondMessageFail = `{"complete2": "hello"}`
+
+const SecondMessagePass = `{"complete2": true}`
