@@ -6,89 +6,87 @@ import (
 	"strings"
 
 	"github.com/alecthomas/jsonschema"
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/protoc-gen-go/descriptor"
 	"github.com/iancoleman/orderedmap"
 	"github.com/xeipuuv/gojsonschema"
+	"google.golang.org/protobuf/types/descriptorpb"
 )
 
 var (
-	globalPkg = &ProtoPackage{
-		name:     "",
-		parent:   nil,
-		children: make(map[string]*ProtoPackage),
-		types:    make(map[string]*descriptor.DescriptorProto),
-	}
+	globalPkg = newProtoPackage(nil, "")
 
-	wellKnownTypes = map[string]string{
-		"DoubleValue": gojsonschema.TYPE_NUMBER,
-		"FloatValue":  gojsonschema.TYPE_NUMBER,
-		"Int64Value":  gojsonschema.TYPE_STRING,
-		"UInt64Value": gojsonschema.TYPE_STRING,
-		"Int32Value":  gojsonschema.TYPE_INTEGER,
-		"UInt32Value": gojsonschema.TYPE_INTEGER,
-		"BoolValue":   gojsonschema.TYPE_BOOLEAN,
-		"StringValue": gojsonschema.TYPE_STRING,
-		"BytesValue":  gojsonschema.TYPE_STRING,
+	wellKnownTypes = map[string]bool{
+		"BoolValue":   true,
+		"BytesValue":  true,
+		"DoubleValue": true,
+		"Duration":    true,
+		"FloatValue":  true,
+		"Int32Value":  true,
+		"Int64Value":  true,
+		"ListValue":   true,
+		"StringValue": true,
+		"Struct":      true,
+		"UInt32Value": true,
+		"UInt64Value": true,
+		"Value":       true,
 	}
 )
 
-func (c *Converter) registerType(pkgName *string, msg *descriptor.DescriptorProto) {
+func (c *Converter) registerEnum(pkgName string, enum *descriptorpb.EnumDescriptorProto) {
 	pkg := globalPkg
-	if pkgName != nil {
-		for _, node := range strings.Split(*pkgName, ".") {
+	if pkgName != "" {
+		for _, node := range strings.Split(pkgName, ".") {
 			if pkg == globalPkg && node == "" {
 				// Skips leading "."
 				continue
 			}
 			child, ok := pkg.children[node]
 			if !ok {
-				child = &ProtoPackage{
-					name:     pkg.name + "." + node,
-					parent:   pkg,
-					children: make(map[string]*ProtoPackage),
-					types:    make(map[string]*descriptor.DescriptorProto),
-				}
+				child = newProtoPackage(pkg, node)
 				pkg.children[node] = child
 			}
 			pkg = child
 		}
 	}
-	pkg.types[msg.GetName()] = msg
+	pkg.enums[enum.GetName()] = enum
 }
 
-func (c *Converter) relativelyLookupNestedType(desc *descriptor.DescriptorProto, name string) (*descriptor.DescriptorProto, bool) {
-	components := strings.Split(name, ".")
-componentLoop:
-	for _, component := range components {
-		for _, nested := range desc.GetNestedType() {
-			if nested.GetName() == component {
-				desc = nested
-				continue componentLoop
+func (c *Converter) registerType(pkgName string, msgDesc *descriptorpb.DescriptorProto) {
+	pkg := globalPkg
+	if pkgName != "" {
+		for _, node := range strings.Split(pkgName, ".") {
+			if pkg == globalPkg && node == "" {
+				// Skips leading "."
+				continue
 			}
+			child, ok := pkg.children[node]
+			if !ok {
+				child = newProtoPackage(pkg, node)
+				pkg.children[node] = child
+			}
+			pkg = child
 		}
-		c.logger.WithField("component", component).WithField("description", desc.GetName()).Info("no such nested message")
-		return nil, false
 	}
-	return desc, true
+	pkg.types[msgDesc.GetName()] = msgDesc
 }
 
 // Convert a proto "field" (essentially a type-switch with some recursion):
-func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDescriptorProto, msg *descriptor.DescriptorProto) (*jsonschema.Type, error) {
+func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptorpb.FieldDescriptorProto, msgDesc *descriptorpb.DescriptorProto, duplicatedMessages map[*descriptorpb.DescriptorProto]string, messageFlags ConverterFlags) (*jsonschema.Type, error) {
 
 	// Prepare a new jsonschema.Type for our eventual return value:
 	jsonSchemaType := &jsonschema.Type{}
 
 	// Generate a description from src comments (if available)
 	if src := c.sourceInfo.GetField(desc); src != nil {
-		jsonSchemaType.Description = formatDescription(src)
+		jsonSchemaType.Title, jsonSchemaType.Description = c.formatTitleAndDescription("", src)
 	}
 
 	// Switch the types, and pick a JSONSchema equivalent:
 	switch desc.GetType() {
-	case descriptor.FieldDescriptorProto_TYPE_DOUBLE,
-		descriptor.FieldDescriptorProto_TYPE_FLOAT:
-		if c.AllowNullValues {
+
+	// Float32:
+	case descriptorpb.FieldDescriptorProto_TYPE_DOUBLE,
+		descriptorpb.FieldDescriptorProto_TYPE_FLOAT:
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
 				{Type: gojsonschema.TYPE_NUMBER},
@@ -97,12 +95,13 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.Type = gojsonschema.TYPE_NUMBER
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_INT32,
-		descriptor.FieldDescriptorProto_TYPE_UINT32,
-		descriptor.FieldDescriptorProto_TYPE_FIXED32,
-		descriptor.FieldDescriptorProto_TYPE_SFIXED32,
-		descriptor.FieldDescriptorProto_TYPE_SINT32:
-		if c.AllowNullValues {
+	// Int32:
+	case descriptorpb.FieldDescriptorProto_TYPE_INT32,
+		descriptorpb.FieldDescriptorProto_TYPE_UINT32,
+		descriptorpb.FieldDescriptorProto_TYPE_FIXED32,
+		descriptorpb.FieldDescriptorProto_TYPE_SFIXED32,
+		descriptorpb.FieldDescriptorProto_TYPE_SINT32:
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
 				{Type: gojsonschema.TYPE_INTEGER},
@@ -111,56 +110,91 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.Type = gojsonschema.TYPE_INTEGER
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_INT64,
-		descriptor.FieldDescriptorProto_TYPE_UINT64,
-		descriptor.FieldDescriptorProto_TYPE_FIXED64,
-		descriptor.FieldDescriptorProto_TYPE_SFIXED64,
-		descriptor.FieldDescriptorProto_TYPE_SINT64:
-		jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_INTEGER})
-		if !c.DisallowBigIntsAsStrings {
-			jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_STRING})
-		}
-		if c.AllowNullValues {
-			jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_NULL})
+	// Int64:
+	case descriptorpb.FieldDescriptorProto_TYPE_INT64,
+		descriptorpb.FieldDescriptorProto_TYPE_UINT64,
+		descriptorpb.FieldDescriptorProto_TYPE_FIXED64,
+		descriptorpb.FieldDescriptorProto_TYPE_SFIXED64,
+		descriptorpb.FieldDescriptorProto_TYPE_SINT64:
+
+		// As integer:
+		if c.Flags.DisallowBigIntsAsStrings {
+			if messageFlags.AllowNullValues {
+				jsonSchemaType.OneOf = []*jsonschema.Type{
+					{Type: gojsonschema.TYPE_INTEGER},
+					{Type: gojsonschema.TYPE_NULL},
+				}
+			} else {
+				jsonSchemaType.Type = gojsonschema.TYPE_INTEGER
+			}
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_STRING,
-		descriptor.FieldDescriptorProto_TYPE_BYTES:
-		if c.AllowNullValues {
+		// As string:
+		if !c.Flags.DisallowBigIntsAsStrings {
+			if messageFlags.AllowNullValues {
+				jsonSchemaType.OneOf = []*jsonschema.Type{
+					{Type: gojsonschema.TYPE_STRING},
+					{Type: gojsonschema.TYPE_NULL},
+				}
+			} else {
+				jsonSchemaType.Type = gojsonschema.TYPE_STRING
+			}
+		}
+
+	// String:
+	case descriptorpb.FieldDescriptorProto_TYPE_STRING:
+		stringDef := &jsonschema.Type{Type: gojsonschema.TYPE_STRING}
+
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
-				{Type: gojsonschema.TYPE_STRING},
+				stringDef,
+			}
+		} else {
+			jsonSchemaType.Type = stringDef.Type
+			jsonSchemaType.MinLength = stringDef.MinLength
+			jsonSchemaType.MaxLength = stringDef.MaxLength
+			jsonSchemaType.Pattern = stringDef.Pattern
+		}
+
+	// Bytes:
+	case descriptorpb.FieldDescriptorProto_TYPE_BYTES:
+		if messageFlags.AllowNullValues {
+			jsonSchemaType.OneOf = []*jsonschema.Type{
+				{Type: gojsonschema.TYPE_NULL},
+				{
+					Type:           gojsonschema.TYPE_STRING,
+					Format:         "binary",
+					BinaryEncoding: "base64",
+				},
 			}
 		} else {
 			jsonSchemaType.Type = gojsonschema.TYPE_STRING
+			jsonSchemaType.Format = "binary"
+			jsonSchemaType.BinaryEncoding = "base64"
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_ENUM:
-		jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_STRING})
-		jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_INTEGER})
-		if c.AllowNullValues {
-			jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_NULL})
+	// ENUM:
+	case descriptorpb.FieldDescriptorProto_TYPE_ENUM:
+
+		// Go through all the enums we have, see if we can match any to this field.
+		fullEnumIdentifier := strings.TrimPrefix(desc.GetTypeName(), ".")
+		matchedEnum, _, ok := c.lookupEnum(curPkg, fullEnumIdentifier)
+		if !ok {
+			return nil, fmt.Errorf("unable to resolve enum type: %s", desc.GetType().String())
 		}
 
-		// Go through all the enums we have, see if we can match any to this field by name:
-		for _, enumDescriptor := range msg.GetEnumType() {
-
-			// Each one has several values:
-			for _, enumValue := range enumDescriptor.Value {
-
-				// Figure out the entire name of this field:
-				fullFieldName := fmt.Sprintf(".%v.%v", *msg.Name, *enumDescriptor.Name)
-
-				// If we find ENUM values for this field then put them into the JSONSchema list of allowed ENUM values:
-				if strings.HasSuffix(desc.GetTypeName(), fullFieldName) {
-					jsonSchemaType.Enum = append(jsonSchemaType.Enum, enumValue.Name)
-					jsonSchemaType.Enum = append(jsonSchemaType.Enum, enumValue.Number)
-				}
-			}
+		// We already have a converter for standalone ENUMs, so just use that:
+		enumSchema, err := c.convertEnumType(matchedEnum, messageFlags)
+		if err != nil {
+			return nil, err
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_BOOL:
-		if c.AllowNullValues {
+		jsonSchemaType = &enumSchema
+
+	// Bool:
+	case descriptorpb.FieldDescriptorProto_TYPE_BOOL:
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
 				{Type: gojsonschema.TYPE_BOOLEAN},
@@ -169,22 +203,40 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.Type = gojsonschema.TYPE_BOOLEAN
 		}
 
-	case descriptor.FieldDescriptorProto_TYPE_GROUP,
-		descriptor.FieldDescriptorProto_TYPE_MESSAGE:
-		jsonSchemaType.Type = gojsonschema.TYPE_OBJECT
-		if desc.GetLabel() == descriptor.FieldDescriptorProto_LABEL_OPTIONAL {
+	// Group (object):
+	case descriptorpb.FieldDescriptorProto_TYPE_GROUP, descriptorpb.FieldDescriptorProto_TYPE_MESSAGE:
+
+		switch desc.GetTypeName() {
+		// Make sure that durations match a particular string pattern (eg 3.4s):
+		case ".google.protobuf.Duration":
+			jsonSchemaType.Type = gojsonschema.TYPE_STRING
+			jsonSchemaType.Format = "regex"
+			jsonSchemaType.Pattern = `^([0-9]+\.?[0-9]*|\.[0-9]+)s$`
+		case ".google.protobuf.Timestamp":
+			jsonSchemaType.Type = gojsonschema.TYPE_STRING
+			jsonSchemaType.Format = "date-time"
+		case ".google.protobuf.Value", ".google.protobuf.Struct":
+			jsonSchemaType.Type = gojsonschema.TYPE_OBJECT
 			jsonSchemaType.AdditionalProperties = []byte("true")
-		}
-		if desc.GetLabel() == descriptor.FieldDescriptorProto_LABEL_REQUIRED {
-			jsonSchemaType.AdditionalProperties = []byte("false")
+		default:
+			jsonSchemaType.Type = gojsonschema.TYPE_OBJECT
+			if desc.GetLabel() == descriptorpb.FieldDescriptorProto_LABEL_OPTIONAL {
+				jsonSchemaType.AdditionalProperties = []byte("true")
+			}
+			if desc.GetLabel() == descriptorpb.FieldDescriptorProto_LABEL_REQUIRED {
+				jsonSchemaType.AdditionalProperties = []byte("false")
+			}
+			if messageFlags.DisallowAdditionalProperties {
+				jsonSchemaType.AdditionalProperties = []byte("false")
+			}
 		}
 
 	default:
 		return nil, fmt.Errorf("unrecognized field type: %s", desc.GetType().String())
 	}
 
-	// Recurse array of primitive types:
-	if desc.GetLabel() == descriptor.FieldDescriptorProto_LABEL_REPEATED && jsonSchemaType.Type != gojsonschema.TYPE_OBJECT {
+	// Recurse basic array:
+	if desc.GetLabel() == descriptorpb.FieldDescriptorProto_LABEL_REPEATED && jsonSchemaType.Type != gojsonschema.TYPE_OBJECT {
 		jsonSchemaType.Items = &jsonschema.Type{}
 
 		if len(jsonSchemaType.Enum) > 0 {
@@ -196,7 +248,7 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.Items.OneOf = jsonSchemaType.OneOf
 		}
 
-		if c.AllowNullValues {
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
 				{Type: gojsonschema.TYPE_ARRAY},
@@ -205,7 +257,6 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.Type = gojsonschema.TYPE_ARRAY
 			jsonSchemaType.OneOf = []*jsonschema.Type{}
 		}
-
 		return jsonSchemaType, nil
 	}
 
@@ -218,8 +269,7 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 		}
 
 		// Recurse the recordType:
-		recursedJSONSchemaType, err := c.convertMessageType(curPkg, recordType, pkgName)
-
+		recursedJSONSchemaType, err := c.recursiveConvertMessageType(curPkg, recordType, pkgName, duplicatedMessages, false)
 		if err != nil {
 			return nil, err
 		}
@@ -231,14 +281,16 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 		case recordType.Options.GetMapEntry():
 			c.logger.
 				WithField("field_name", recordType.GetName()).
-				WithField("msg_name", *msg.Name).
+				WithField("msgDesc_name", *msgDesc.Name).
 				Tracef("Is a map")
-			// Make sure we have a "value":
+
 			if recursedJSONSchemaType.Properties == nil {
-				return nil, fmt.Errorf("Unable to find 'value' property of MAP type")
+				return nil, fmt.Errorf("Unable to find properties of MAP type")
 			}
-			value, ok := recursedJSONSchemaType.Properties.Get("value")
-			if !ok {
+
+			// Make sure we have a "value":
+			value, valuePresent := recursedJSONSchemaType.Properties.Get("value")
+			if !valuePresent {
 				return nil, fmt.Errorf("Unable to find 'value' property of MAP type")
 			}
 
@@ -250,65 +302,228 @@ func (c *Converter) convertField(curPkg *ProtoPackage, desc *descriptor.FieldDes
 			jsonSchemaType.AdditionalProperties = additionalPropertiesJSON
 
 		// Arrays:
-		case desc.GetLabel() == descriptor.FieldDescriptorProto_LABEL_REPEATED:
-			jsonSchemaType.Items = &recursedJSONSchemaType
+		case desc.GetLabel() == descriptorpb.FieldDescriptorProto_LABEL_REPEATED:
+			jsonSchemaType.Items = recursedJSONSchemaType
 			jsonSchemaType.Type = gojsonschema.TYPE_ARRAY
 
-		// Objects:
-		default:
-			isPrimitive := true
-			for _, t := range recursedJSONSchemaType.OneOf {
-				if t.Type == gojsonschema.TYPE_OBJECT || t.Type == gojsonschema.TYPE_ARRAY {
-					isPrimitive = false
+			// Build up the list of required fields:
+			if messageFlags.AllFieldsRequired && len(recursedJSONSchemaType.OneOf) == 0 && recursedJSONSchemaType.Properties != nil {
+				for _, property := range recursedJSONSchemaType.Properties.Keys() {
+					jsonSchemaType.Items.Required = append(jsonSchemaType.Items.Required, property)
 				}
 			}
-			if recursedJSONSchemaType.OneOf != nil && isPrimitive {
-				jsonSchemaType.AdditionalProperties = nil
-				jsonSchemaType.Type = ""
-				jsonSchemaType.OneOf = recursedJSONSchemaType.OneOf
-			} else {
-				jsonSchemaType.Properties = recursedJSONSchemaType.Properties
+			jsonSchemaType.Items.Required = dedupe(jsonSchemaType.Items.Required)
+
+		// Not maps, not arrays:
+		default:
+
+			// If we've got optional types then just take those:
+			if recursedJSONSchemaType.OneOf != nil {
+				return recursedJSONSchemaType, nil
+			}
+
+			// If we're not an object then set the type from whatever we recursed:
+			if recursedJSONSchemaType.Type != gojsonschema.TYPE_OBJECT {
+				jsonSchemaType.Type = recursedJSONSchemaType.Type
+			}
+
+			// Assume the attrbutes of the recursed value:
+			jsonSchemaType.Properties = recursedJSONSchemaType.Properties
+			jsonSchemaType.Ref = recursedJSONSchemaType.Ref
+			jsonSchemaType.Required = recursedJSONSchemaType.Required
+
+			// Build up the list of required fields:
+			if messageFlags.AllFieldsRequired && len(recursedJSONSchemaType.OneOf) == 0 && recursedJSONSchemaType.Properties != nil {
+				for _, property := range recursedJSONSchemaType.Properties.Keys() {
+					jsonSchemaType.Required = append(jsonSchemaType.Required, property)
+				}
 			}
 		}
 
-		// Optionally allow NULL values, if not already nullable
-		if c.AllowNullValues && jsonSchemaType.OneOf == nil {
+		// Optionally allow NULL values:
+		if messageFlags.AllowNullValues {
 			jsonSchemaType.OneOf = []*jsonschema.Type{
 				{Type: gojsonschema.TYPE_NULL},
-				{Type: jsonSchemaType.Type},
+				{Type: jsonSchemaType.Type, Items: jsonSchemaType.Items},
 			}
 			jsonSchemaType.Type = ""
+			jsonSchemaType.Items = nil
 		}
 	}
+
+	jsonSchemaType.Required = dedupe(jsonSchemaType.Required)
 
 	return jsonSchemaType, nil
 }
 
 // Converts a proto "MESSAGE" into a JSON-Schema:
-func (c *Converter) convertMessageType(curPkg *ProtoPackage, msg *descriptor.DescriptorProto, pkgName string) (jsonschema.Type, error) {
-	if msg.Name != nil && pkgName == ".google.protobuf" {
-		if jsonType, ok := wellKnownTypes[*msg.Name]; ok {
-			schema := jsonschema.Type{
-				OneOf: []*jsonschema.Type{
-					{Type: gojsonschema.TYPE_NULL},
-					{Type: jsonType},
-				},
-			}
-			return schema, nil
+func (c *Converter) convertMessageType(curPkg *ProtoPackage, msgDesc *descriptorpb.DescriptorProto) (*jsonschema.Schema, error) {
+
+	// Get a list of any nested messages in our schema:
+	duplicatedMessages, err := c.findNestedMessages(curPkg, msgDesc)
+	if err != nil {
+		return nil, err
+	}
+
+	// Build up a list of JSONSchema type definitions for every message:
+	definitions := jsonschema.Definitions{}
+	for refmsgDesc, nameWithPackage := range duplicatedMessages {
+		var typeName string
+		if c.Flags.TypeNamesWithNoPackage {
+			typeName = refmsgDesc.GetName()
+		} else {
+			typeName = nameWithPackage
+		}
+		refType, err := c.recursiveConvertMessageType(curPkg, refmsgDesc, "", duplicatedMessages, true)
+		if err != nil {
+			return nil, err
+		}
+
+		// Add the schema to our definitions:
+		definitions[typeName] = refType
+	}
+
+	// Put together a JSON schema with our discovered definitions, and a $ref for the root type:
+	newJSONSchema := &jsonschema.Schema{
+		Type: &jsonschema.Type{
+			Ref:     fmt.Sprintf("%s%s", c.refPrefix, msgDesc.GetName()),
+			Version: c.schemaVersion,
+		},
+		Definitions: definitions,
+	}
+
+	return newJSONSchema, nil
+}
+
+// findNestedMessages takes a message, and returns a map mapping pointers to messages nested within it:
+// these messages become definitions which can be referenced (instead of repeating them every time they're used)
+func (c *Converter) findNestedMessages(curPkg *ProtoPackage, msgDesc *descriptorpb.DescriptorProto) (map[*descriptorpb.DescriptorProto]string, error) {
+
+	// Get a list of all nested messages, and how often they occur:
+	nestedMessages := make(map[*descriptorpb.DescriptorProto]string)
+	if err := c.recursiveFindNestedMessages(curPkg, msgDesc, msgDesc.GetName(), nestedMessages); err != nil {
+		return nil, err
+	}
+
+	// Now filter them:
+	result := make(map[*descriptorpb.DescriptorProto]string)
+	for message, messageName := range nestedMessages {
+		if !message.GetOptions().GetMapEntry() && !strings.HasPrefix(messageName, ".google.protobuf.") {
+			result[message] = strings.TrimLeft(messageName, ".")
 		}
 	}
 
-	// Prepare a new jsonschema:
-	jsonSchemaType := jsonschema.Type{
-		Version: jsonschema.Version,
+	return result, nil
+}
+
+func (c *Converter) recursiveFindNestedMessages(curPkg *ProtoPackage, msgDesc *descriptorpb.DescriptorProto, typeName string, nestedMessages map[*descriptorpb.DescriptorProto]string) error {
+	if _, present := nestedMessages[msgDesc]; present {
+		return nil
 	}
+	nestedMessages[msgDesc] = typeName
+
+	for _, desc := range msgDesc.GetField() {
+		descType := desc.GetType()
+		if descType != descriptorpb.FieldDescriptorProto_TYPE_MESSAGE && descType != descriptorpb.FieldDescriptorProto_TYPE_GROUP {
+			// no nested messages
+			continue
+		}
+
+		typeName := desc.GetTypeName()
+		recordType, _, ok := c.lookupType(curPkg, typeName)
+		if !ok {
+			return fmt.Errorf("no such message type named %s", typeName)
+		}
+		if err := c.recursiveFindNestedMessages(curPkg, recordType, typeName, nestedMessages); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (c *Converter) recursiveConvertMessageType(curPkg *ProtoPackage, msgDesc *descriptorpb.DescriptorProto, pkgName string, duplicatedMessages map[*descriptorpb.DescriptorProto]string, ignoreDuplicatedMessages bool) (*jsonschema.Type, error) {
+
+	// Prepare a new jsonschema:
+	jsonSchemaType := new(jsonschema.Type)
+
+	// Set some per-message flags from config and options:
+	messageFlags := c.Flags
+
 	// Generate a description from src comments (if available)
-	if src := c.sourceInfo.GetMessage(msg); src != nil {
-		jsonSchemaType.Description = formatDescription(src)
+	if src := c.sourceInfo.GetMessage(msgDesc); src != nil {
+		jsonSchemaType.Title, jsonSchemaType.Description = c.formatTitleAndDescription(msgDesc.GetName(), src)
+	}
+
+	// Handle google's well-known types:
+	if name := msgDesc.GetName(); wellKnownTypes[name] && pkgName == ".google.protobuf" {
+		switch name {
+		case "DoubleValue", "FloatValue":
+			jsonSchemaType.Type = gojsonschema.TYPE_NUMBER
+		case "Int32Value", "UInt32Value":
+			jsonSchemaType.Type = gojsonschema.TYPE_INTEGER
+		case "Int64Value", "UInt64Value":
+			// BigInt as ints
+			if messageFlags.DisallowBigIntsAsStrings {
+				jsonSchemaType.Type = gojsonschema.TYPE_INTEGER
+			} else {
+
+				// BigInt as strings
+				jsonSchemaType.Type = gojsonschema.TYPE_STRING
+			}
+
+		case "BoolValue":
+			jsonSchemaType.Type = gojsonschema.TYPE_BOOLEAN
+		case "BytesValue", "StringValue":
+			jsonSchemaType.Type = gojsonschema.TYPE_STRING
+		case "Value":
+			jsonSchemaType.OneOf = []*jsonschema.Type{
+				{Type: gojsonschema.TYPE_ARRAY},
+				{Type: gojsonschema.TYPE_BOOLEAN},
+				{Type: gojsonschema.TYPE_NUMBER},
+				{Type: gojsonschema.TYPE_OBJECT},
+				{Type: gojsonschema.TYPE_STRING},
+			}
+			// jsonSchemaType.AdditionalProperties = []byte("true")
+		case "Duration":
+			jsonSchemaType.Type = gojsonschema.TYPE_STRING
+		case "Struct":
+			jsonSchemaType.Type = gojsonschema.TYPE_OBJECT
+			// jsonSchemaType.AdditionalProperties = []byte("true")
+		case "ListValue":
+			jsonSchemaType.Type = gojsonschema.TYPE_ARRAY
+		}
+
+		// If we're allowing nulls then prepare a OneOf:
+		if messageFlags.AllowNullValues {
+			jsonSchemaType.OneOf = append(jsonSchemaType.OneOf, &jsonschema.Type{Type: gojsonschema.TYPE_NULL}, &jsonschema.Type{Type: jsonSchemaType.Type})
+			// and clear the Type that was previously set.
+			jsonSchemaType.Type = ""
+			return jsonSchemaType, nil
+		}
+
+		// Otherwise just return this simple type:
+		return jsonSchemaType, nil
+	}
+
+	// Set defaults:
+	jsonSchemaType.Properties = orderedmap.New()
+
+	// Look up references:
+	if nameWithPackage, ok := duplicatedMessages[msgDesc]; ok && !ignoreDuplicatedMessages {
+		var typeName string
+		if c.Flags.TypeNamesWithNoPackage {
+			typeName = msgDesc.GetName()
+		} else {
+			typeName = nameWithPackage
+		}
+		return &jsonschema.Type{
+			Ref: fmt.Sprintf("%s%s", c.refPrefix, typeName),
+		}, nil
 	}
 
 	// Optionally allow NULL values:
-	if c.AllowNullValues {
+	if messageFlags.AllowNullValues {
 		jsonSchemaType.OneOf = []*jsonschema.Type{
 			{Type: gojsonschema.TYPE_NULL},
 			{Type: gojsonschema.TYPE_OBJECT},
@@ -318,44 +533,93 @@ func (c *Converter) convertMessageType(curPkg *ProtoPackage, msg *descriptor.Des
 	}
 
 	// disallowAdditionalProperties will prevent validation where extra fields are found (outside of the schema):
-	if c.DisallowAdditionalProperties {
+	if messageFlags.DisallowAdditionalProperties {
 		jsonSchemaType.AdditionalProperties = []byte("false")
 	} else {
 		jsonSchemaType.AdditionalProperties = []byte("true")
 	}
 
-	c.logger.WithField("message_str", proto.MarshalTextString(msg)).Trace("Converting message")
-	for _, fieldDesc := range msg.GetField() {
-		recursedJSONSchemaType, err := c.convertField(curPkg, fieldDesc, msg)
+	c.logger.WithField("message_str", msgDesc.String()).Trace("Converting message")
+	for _, fieldDesc := range msgDesc.GetField() {
+
+		// Convert the field into a JSONSchema type:
+		recursedJSONSchemaType, err := c.convertField(curPkg, fieldDesc, msgDesc, duplicatedMessages, messageFlags)
 		if err != nil {
-			c.logger.WithError(err).WithField("field_name", fieldDesc.GetName()).WithField("message_name", msg.GetName()).Error("Failed to convert field")
-			return jsonSchemaType, err
+			c.logger.WithError(err).WithField("field_name", fieldDesc.GetName()).WithField("message_name", msgDesc.GetName()).Error("Failed to convert field")
+			return nil, err
 		}
-		c.logger.WithField("field_name", fieldDesc.GetName()).WithField("type", recursedJSONSchemaType.Type).Debug("Converted field")
-		if jsonSchemaType.Properties == nil {
-			jsonSchemaType.Properties = orderedmap.New()
+		c.logger.WithField("field_name", fieldDesc.GetName()).WithField("type", recursedJSONSchemaType.Type).Trace("Converted field")
+
+		// If this field is part of a OneOf declaration then build that here:
+		if c.Flags.EnforceOneOf && fieldDesc.OneofIndex != nil && !fieldDesc.GetProto3Optional() {
+			for {
+				if *fieldDesc.OneofIndex < int32(len(jsonSchemaType.AllOf)) {
+					break
+				}
+				var notAnyOf = &jsonschema.Type{Not: &jsonschema.Type{AnyOf: []*jsonschema.Type{}}}
+				jsonSchemaType.AllOf = append(jsonSchemaType.AllOf, &jsonschema.Type{OneOf: []*jsonschema.Type{notAnyOf}})
+			}
+			if c.Flags.UseJSONFieldnamesOnly {
+				jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf = append(jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf, &jsonschema.Type{Required: []string{fieldDesc.GetJsonName()}})
+				jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf[0].Not.AnyOf = append(jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf[0].Not.AnyOf, &jsonschema.Type{Required: []string{fieldDesc.GetJsonName()}})
+			} else {
+				jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf = append(jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf, &jsonschema.Type{Required: []string{fieldDesc.GetName()}})
+				jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf[0].Not.AnyOf = append(jsonSchemaType.AllOf[*fieldDesc.OneofIndex].OneOf[0].Not.AnyOf, &jsonschema.Type{Required: []string{fieldDesc.GetName()}})
+			}
 		}
-		jsonSchemaType.Properties.Set(fieldDesc.GetName(), recursedJSONSchemaType)
-		if c.UseProtoAndJSONFieldnames && fieldDesc.GetName() != fieldDesc.GetJsonName() {
+
+		// Figure out which field names we want to use:
+		switch {
+		case c.Flags.UseJSONFieldnamesOnly:
 			jsonSchemaType.Properties.Set(fieldDesc.GetJsonName(), recursedJSONSchemaType)
+		case c.Flags.UseProtoAndJSONFieldNames:
+			jsonSchemaType.Properties.Set(fieldDesc.GetName(), recursedJSONSchemaType)
+			jsonSchemaType.Properties.Set(fieldDesc.GetJsonName(), recursedJSONSchemaType)
+		default:
+			jsonSchemaType.Properties.Set(fieldDesc.GetName(), recursedJSONSchemaType)
+		}
+
+		// Enforce all_fields_required:
+		if messageFlags.AllFieldsRequired {
+			if fieldDesc.OneofIndex == nil && !fieldDesc.GetProto3Optional() {
+				if c.Flags.UseJSONFieldnamesOnly {
+					jsonSchemaType.Required = append(jsonSchemaType.Required, fieldDesc.GetJsonName())
+				} else {
+					jsonSchemaType.Required = append(jsonSchemaType.Required, fieldDesc.GetName())
+				}
+			}
+		}
+
+		// Look for required fields by the proto2 "required" flag:
+		if fieldDesc.GetLabel() == descriptorpb.FieldDescriptorProto_LABEL_REQUIRED && fieldDesc.OneofIndex == nil {
+			if c.Flags.UseJSONFieldnamesOnly {
+				jsonSchemaType.Required = append(jsonSchemaType.Required, fieldDesc.GetJsonName())
+			} else {
+				jsonSchemaType.Required = append(jsonSchemaType.Required, fieldDesc.GetName())
+			}
 		}
 	}
+
+	// Remove empty properties to keep the final output as clean as possible:
+	if len(jsonSchemaType.Properties.Keys()) == 0 {
+		jsonSchemaType.Properties = nil
+	}
+
+	// Dedupe required fields:
+	jsonSchemaType.Required = dedupe(jsonSchemaType.Required)
 
 	return jsonSchemaType, nil
 }
 
-func formatDescription(sl *descriptor.SourceCodeInfo_Location) string {
-	var lines []string
-	for _, str := range sl.GetLeadingDetachedComments() {
-		if s := strings.TrimSpace(str); s != "" {
-			lines = append(lines, s)
+func dedupe(inputStrings []string) []string {
+	appended := make(map[string]bool)
+	outputStrings := []string{}
+
+	for _, inputString := range inputStrings {
+		if !appended[inputString] {
+			outputStrings = append(outputStrings, inputString)
+			appended[inputString] = true
 		}
 	}
-	if s := strings.TrimSpace(sl.GetLeadingComments()); s != "" {
-		lines = append(lines, s)
-	}
-	if s := strings.TrimSpace(sl.GetTrailingComments()); s != "" {
-		lines = append(lines, s)
-	}
-	return strings.Join(lines, "\n\n")
+	return outputStrings
 }
